requirements.txt
alabaster
Babel
carbontracker
certifi
charset-normalizer
click
contourpy
coverage
cycler
docutils
filelock
flake8
fonttools
fsspec
idna
imageio
imagesize
Jinja2
kiwisolver
lazy_loader
MarkupSafe
matplotlib
mccabe
mpmath
networkx
numpy
packaging
pandas
pillow
pycodestyle
pyflakes
Pygments
pyparsing
python-dateutil
python-dotenv
pytz
requests
scikit-image
scikit-learn
scipy
seaborn
setuptools
shap
six
snowballstemmer
Sphinx
sphinxcontrib-applehelp
sphinxcontrib-devhelp
sphinxcontrib-htmlhelp
sphinxcontrib-jsmath
sphinxcontrib-qthelp
sphinxcontrib-serializinghtml
sympy
tensorboard
tensorflow
tifffile
tqdm
typing_extensions
tzdata
urllib3
wheel
opencv-python
torcheval
-e .
